+++ html_lore/server/__init__.py
@@ -1,0 +1,1 @@
+"""HTMlore server namespace."""